# app-contactos/swagger-response.json
{"openapi":"3.1.0","info":{"title":"API de Contactos","description":"Documentação para a API de gestão de contactos","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Utilizadores","description":"Gestão de utilizadores"},{"name":"Contactos","description":"Gestão de contactos associados a utilizadores"}],"paths":{"/api/utilizadores/{id}":{"get":{"tags":["Utilizadores"],"summary":"Obter utilizador por ID","description":"Busca um utilizador específico pelo seu ID","operationId":"getUtilizadorById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Utilizador encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}}},"404":{"description":"Utilizador não encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}}}}},"put":{"tags":["Utilizadores"],"summary":"Atualizar utilizador","description":"Atualiza os dados de um utilizador existente","operationId":"updateUtilizador","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}},"required":true},"responses":{"200":{"description":"Utilizador atualizado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}}},"404":{"description":"Utilizador não encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}}},"400":{"description":"Dados inválidos ou email já existente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}}}}},"delete":{"tags":["Utilizadores"],"summary":"Eliminar utilizador","description":"Remove um utilizador do sistema","operationId":"deleteUtilizador","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Utilizador eliminado com sucesso"},"404":{"description":"Utilizador não encontrado"}}}},"/api/contactos/{id}":{"get":{"tags":["Contactos"],"summary":"Obter contacto por ID","description":"Busca um contacto específico pelo seu ID","operationId":"getContactoById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Contacto encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContactoDTO"}}}},"404":{"description":"Contacto não encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContactoDTO"}}}}}},"put":{"tags":["Contactos"],"summary":"Atualizar contacto","description":"Atualiza os dados de um contacto existente","operationId":"updateContacto","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactoDTO"}}},"required":true},"responses":{"200":{"description":"Contacto atualizado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContactoDTO"}}}},"404":{"description":"Contacto não encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContactoDTO"}}}},"400":{"description":"Dados inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContactoDTO"}}}}}},"delete":{"tags":["Contactos"],"summary":"Eliminar contacto","description":"Remove um contacto do sistema","operationId":"deleteContacto","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Contacto eliminado com sucesso"},"404":{"description":"Contacto não encontrado"}}}},"/api/utilizadores":{"get":{"tags":["Utilizadores"],"summary":"Listar todos os utilizadores","description":"Retorna todos os utilizadores existentes","operationId":"getAllUtilizadores","responses":{"200":{"description":"Lista de utilizadores","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UtilizadorDTO"}}}}}}},"post":{"tags":["Utilizadores"],"summary":"Criar utilizador","description":"Regista um novo utilizador","operationId":"createUtilizador","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}},"required":true},"responses":{"201":{"description":"Utilizador criado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}}},"400":{"description":"Dados inválidos ou email já existente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}}}}}},"/api/contactos":{"get":{"tags":["Contactos"],"summary":"Listar todos os contactos","description":"Retorna todos os contactos existentes","operationId":"getAllContactos","responses":{"200":{"description":"Lista de contactos","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ContactoDTO"}}}}}}},"post":{"tags":["Contactos"],"summary":"Criar contacto","description":"Associa um contacto a um utilizador existente","operationId":"createContacto","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactoDTO"}}},"required":true},"responses":{"201":{"description":"Contacto criado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContactoDTO"}}}},"400":{"description":"Dados inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContactoDTO"}}}},"404":{"description":"Utilizador não encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContactoDTO"}}}}}}},"/api/utilizadores/{id}/exists":{"get":{"tags":["Utilizadores"],"summary":"Verificar existência de utilizador por ID","description":"Verifica se um utilizador existe pelo seu ID","operationId":"existsById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Status de existência do utilizador","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/utilizadores/email":{"get":{"tags":["Utilizadores"],"summary":"Obter utilizador por email","description":"Busca um utilizador específico pelo seu email","operationId":"getUtilizadorByEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Utilizador encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}}},"404":{"description":"Utilizador não encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UtilizadorDTO"}}}}}}},"/api/utilizadores/email/exists":{"get":{"tags":["Utilizadores"],"summary":"Verificar existência de utilizador por email","description":"Verifica se um utilizador existe pelo seu email","operationId":"existsByEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Status de existência do utilizador","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/contactos/{id}/exists":{"get":{"tags":["Contactos"],"summary":"Verificar existência de contacto","description":"Verifica se um contacto existe pelo seu ID","operationId":"existsById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Status de existência do contacto","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/contactos/utilizador/{utilizadorId}":{"get":{"tags":["Contactos"],"summary":"Listar contactos por utilizador","description":"Retorna todos os contactos de um utilizador específico","operationId":"getContactosByUtilizadorId","parameters":[{"name":"utilizadorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Lista de contactos do utilizador","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ContactoDTO"}}}}}}}}},"components":{"schemas":{"UtilizadorDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string","maxLength":100,"minLength":2},"email":{"type":"string","maxLength":150,"minLength":0}},"required":["email","nome"]},"ContactoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string","maxLength":100,"minLength":2},"telefone":{"type":"string","maxLength":20,"minLength":7,"pattern":"^[+]?[0-9\\s()-]+$"},"utilizadorId":{"type":"integer","format":"int64"}},"required":["nome","utilizadorId"]}}}}